Refactor stateful checkbox props - RSC-1260
